move offsets to offsets32.c
